Missed adding the type cast to avoid the overflow

+++ c/nla-digbench/prod4br-ll.c
@@ -30,7 +30,7 @@ int main() {
     q = 0;
 
     while (1) {
-        __VERIFIER_assert(q + a * b * p == x * y);
+        __VERIFIER_assert(q + a * b * p == (long long) x * y);
 
         if (!(a != 0 && b != 0))
             break;
@@ -52,7 +52,7 @@ int main() {
         }
     }
 
-    __VERIFIER_assert(q == x * y);
+    __VERIFIER_assert(q == (long long) x * y);
     __VERIFIER_assert(a * b == 0);
     return 0;
 }
